+++ Application/ResearchDataManagementPlatform/Program.cs
@@ -10,6 +10,7 @@ using System.Linq;
 using System.Runtime.InteropServices;
 using CommandLine;
 using Rdmp.Core.Startup;
+using Rdmp.UI;
 using Rdmp.UI.TestsAndSetup;
 using ReusableLibraryCode;
 
@@ -37,20 +38,18 @@ namespace ResearchDataManagementPlatform
             }
 
             UsefulStuff.GetParser()
-                       .ParseArguments<ResearchDataManagementPlatformOptions>(args.Except(new[] { "--squirrel-firstrun" }))
+                       .ParseArguments<ResearchDataManagementPlatformOptions>(args)
                        .MapResult(RunApp, err => -1);
         }
 
         private static object RunApp(ResearchDataManagementPlatformOptions arg)
         {
             arg.PopulateConnectionStringsFromYamlIfMissing();
-            arg.GetConnectionStrings(out var c, out var d);
 
             RDMPBootStrapper<RDMPMainForm> bootStrapper =
                 new RDMPBootStrapper<RDMPMainForm>(
                     new EnvironmentInfo(PluginFolders.Main | PluginFolders.Windows),
-                    c?.ConnectionString,
-                    d?.ConnectionString);
+                    arg);
 
             bootStrapper.Show(false);
             return 0;